Rewrite code to split it into separate controllers
